+++ build4.sh
@@ -12,4 +12,8 @@ DOCKER_BUILDKIT=1 docker build --build-arg OPENAI_API_KEY=$GPT4_KEY \
              -t alicetyan-gpt4:latest \
              .
 
-docker run --rm alicetyan-gpt4:latest env
+if [ "$PACKAGE" = '1' ]; then
+  docker run --rm alicetyan-gpt4:latest env
+  docker save alicetyan-gpt4:latest -o alicetyan-gpt4.tar
+  gzip alicetyan-gpt4.tar
+fi
